+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New algorithm to pretty print which fits more on a line if possible
 
+### Added
+
+- Added soft_wrap flag to Console.print
+
 ## [5.2.1] - 2020-08-19
 
 ### Fixed